refactor: budgets 폴더 내 컴포넌트 성격별 구조화 #81
